src/app/modules/core/components/sidenav/sidenav.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';
import { RouterTestingModule } from '@angular/router/testing';

import { SidenavComponent } from './sidenav.component';

describe('SidenavComponent', () => {
  let component: SidenavComponent;
  let fixture: ComponentFixture<SidenavComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [SidenavComponent],
      imports: [RouterTestingModule]
    }).compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(SidenavComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('signout()', () => {
    component.logout.subscribe(event => {
      expect(event).toBe(undefined);
    });

    component.onSignOut();
  });

  it('onClose()', () => {
    component.closed.subscribe(event => {
      expect(event).toBe(undefined);
    });

    component.onClose();
  });
});
